requirements.txt
qrcode[pil]
pillow
openpyxl
pandas
